Move cloud provider to its own crate
