+++ crates/language_models_cloud/LICENSE-GPL
@@ -1,0 +1,1 @@
+../../LICENSE-GPL
